pyproject.toml
[tool.poetry]
name = "launcher-app"
version = "0.1.0"
description = ""
authors = ["Yakubov, Sergey <yakubovs@ornl.gov>", "Cage, Gregory <cagege@ornl.gov>", "Duggan, John <dugganjw@ornl.gov>"]
readme = "README.md"
license = "MIT"
packages = [{include = "launcher_app"}]
keywords = ["NDIP", "Nova", "python", "dashboard"]

[tool.poetry.dependencies]
python = "^3.10"
bioblend = "^1.3.0"
cryptography = "^43.0.0"
django = "<6"
django-extensions = "^3.2.3"
pyjwt = "^2.9.0"
requests = "^2.32.3"
requests-oauthlib = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  ".", "launcher_app"
]
testpaths = ["tests"]
python_files = ["test*.py"]
norecursedirs = [".git", "tmp*", "_tmp*", "__pycache__"]

[tool.ruff]
line-length = 80
exclude = ["launcher_app/migrations", "settings.py"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "N",  # PEP8 naming convetions
    "D"  # pydocstyle
]
ignore = [
    "C901",  # too complex
    "D102", # Missing docstring in public function
    "D103", # Missing docstring in public method
    "D401"  # imperative mood

]

[tool.ruff.lint.extend-per-file-ignores]
'__init__.py' = ['D104'] # Missing docstring in public package



[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
disallow_untyped_defs = true
disable_error_code = ["attr-defined"]
ignore_errors = false
exclude = ["launcher_app/migrations", "settings.py"]

[tool.coverage.report]
omit = [
    '*__init__*', # __init__ files should just re-export other classes and functions
]

[tool.coverage.run]
command_line = "-m pytest --junit-xml=reports/junit.xml"
data_file = "reports/.coverage"

[tool.poetry.dev-dependencies]
mypy = "^1.10.0"
pre-commit = "^2.20.0"
coverage = "^6.4.3"
pytest = "*"
ruff = "^0.6.2"
types-requests = "*"
